#111 Stops the French ban communism event if banned through decision
The "Ban Communism" decision doesn't add a flag that this event can use in its trigger. Instead, it adds a variable to the communism ideology which prevents it from rising. Having a 5% communism requirement for this event should stop France from getting the "Ban Communism" event if it has already banned communism through a decision.