Updated Yaroslavl state and 1936 starter tech for SOV.
Added a Synthetic Refinery in the Yaroslavl state (Soviet) along with -2 oil so it doesn't produce oil right away. Also added Synthetic Oil Experiments as a starter tech in 1936. The Soviets built a rubber plant in the region in 1932. All of this was a suggestion by Thinking_waffle on the Suggestions chat on discord!