Issue #173 Found the needle in the haystack
After looking for the needle in the haystack, I decided to burn the hay and pull the needle out of the ashes. After deleting almost every other file in the game, I finally narrowed it down to this file, and found that there was an if statement being used incorrectly (which caused the bad trigger). FROM also should have been every country, which caused the bad scope.