Issue #195 Adding GetDate to Czech events
Most of the Czech events seem to be historical descriptions, and I left most of the dates in those alone. However, when an event uses a historical date as the date an event is fired, or a focus is taken, that date is replaced with GetDate. This new file uses a total of 13 GetDate's.