Fixes a CTD caused by forming Vichy France
Someone who plays R56 was nice enough to upload his save game, and I found the source of his crash. The event where France surrenders and becomes Vichy was trying to remove a leader that France didn't have. Adding checks to all the leader removals seems to fix this CTD (although the game lags for a few seconds when Vichy is formed)