add_claim_by = UKR
Considering this is a state with a sizeable amount of Ukrainians and is now administered by UKR, I would suggest adding this claim by them. I, however, do not know if you can add a claim for a country with no territory, but I believe it should work.

I just want to put this out there.