Polishing german focus tree.
Added a statusquo sanity script, what only triggers when the target is in your faction or it is you rsubject.
This prevents the instant bypass of the Prepare the Sealion focus, what previously was impossibile to have it.

U boat effort focus now gives universal subtech bonus, instead of vanilla ones, and unlocks the crusier submarine.