Minor optimizations and bugfixes Chile (#1379)
Hi. Just a quick contribution for some oversights I noticed while playing.

Streamlined allowed_civil_war section on r56_CHL_a_truly_radical_party, r56_CHL_the_popular_front.
CHL_anti_capitalist_idea used local_building_slots_factor rather than global_building_slots_factor, resulting in no effect.
CHL_communist_ban had missing OR on the allowed_civil_war section, making it impossible to fulfill. (You can't be 2 ideologies simultaneously)
Removed redundant AND in CHL_volodia_teitelboim (Its implicit)

The available section of CHL_spread_of_fascist reads: "If Chile it is not a puppet OR it is not a subject" AFAIK, you can't be both simultaneously, so it's always true. Changed so that if either is true, then the idea dies.
The cancel section of CHL_popular_front_frail was always true. Changed so the idea dies if Chile is not independent or fascist. It's gained while neutral, so that part was removed.
CHL_anti_american_idea is only canceled if Chile is not independent and democratic. I won't change it as it may be intentional (cancel only if USA puppets Chile).