Bugfixes and small optimisations ETH focus tree.
The focus ETH_modern_airforce had a duplicated available section.
Rewritten focus ETH_reclaim_aksum to survive in case of unexpected owners of Adan/Yemen.
Small modification to ETH_death_to_saudis to account for the changes on ETH_reclaim_aksum. 
Removed redundant "If" from ETH_red_sea_focus. It's already required, so it has to be true.
Made a little check to ETH_demand_british_somalia to avoid 2 different conquest war goals against the same country.
The focus ETH_demand_solomon_kingdom had some misplaced }. This resulted in a bunch of errors and more importantly, the focus didn't work if ENG wasn't the owner of the Levant.
Formatting.
Removed redundant "AND" from some sections.
Added bypass to the Christmas Offensive and National Mobilization in case the war with Italy ended for some bizarre reason before those focus were taken, mostly cause all the conquest focus are gated behind these. There's still an issue with the support focuses (war with Italy required), but it's a start.

Not fixed is the missing localization for the idea ETH_the_long_lance, given by the focus "The Long Lance".