Fixing MCK file, Take 2
So, editing localization files directly from GitHub can change the format of the file. In specific, it changed the codification from UTF-8-BOM to UTF-8 without BOM. Not sure if it always happens, but its probably why the first line was detected as "changed" despite looking identical.
Uploading the file with the modifications avoids that issue, but I'm unsure as to the final codification after solving the ensuing conflict. It also makes seeing the modification messy.
I'm testing editing the file from "GitHub Desktop". This should work, but I never used it before so... finger crossed.

Regardless, I'll be more careful with future pull requests.