Event cleanup and portrait capitalization
Added relevant capitalisation to Carl Gustaf Fleischer and Josef Terboven's portraits, moved a Dano-Norwegian deal follow up event next to the relevant events, moved the Fennoscandian one out of the way. Also added updated the tooltips for earlier changes done in the last merged commit. The event file should be cleaner to look at and work with from now on.