Singapore addition, Indonesia and Sarawak fix
Added Singapore, minor fixes with Sarawak, and added NU (Nahdlatul Ulama) as a post-independence, neutrality party with it's founder, Hasyim Asy'ari as it's leader. Changed Indonesian flags for fascism, Majapahit, and occupation for historical accuracy. Replaced fictional communist flag with another one that combines the PKI election symbol with the regular Indonesian flag.