Update raj_on_actions.txt
Corrected a scope to what I believe was intended. What was here before was ignored, as it was checking the tag of the province itself rather than from any of the 2 nations involved. It also gave an error, that happens to be the reason I noticed in the first place.