Indonesia content
Added marhaenism as a subideology of both fascism and neutrality and Islamic democracy as a subideology of democracy.

Added an event to decide on the post-independent neutral and democratic parties.

Made minor fix to the focus "Democratic Transition". You can't have Suharto without his party.

Added more generals as a reward to "expand the officer corps" plus one naval commander.

Gave generic portraits to Singapore's leaders as a temporary placeholder.