gfx changes in Asia
Added namelists to Singapore and Malaysia.

Added a monarchist option for The First Sila and one cosmetic tag for each except the historical one. Note that each cosmetic tag is named after the neutral party, and oppositions of other ideologies would naturally be of opposite nature (religious opposition for a secular ruling party and vice versa).

New flags for both Koreas, as well as localization changes.