Korea etc
Split Korea into two, made characters for both, cosmetic tag and (non-functional) decision for a unified Korea (Phantom offered to help with this after PR is up). Added Juche as a communist sub ideology without gfx (so it uses regular communism gfx for now).

Some characters are incomplete, most lacking a portrait. Requests will be made after the last was completed so as not to burden the artists.

Addition of advisor pics for Bhutan, added leader pics for TNG and SIK.