Order
Changed the order so that Puyi, as the constitutional monarch, is put on the top of the list of government officials, close to Zhang Haipeng as a prime minister.

Important Note:

There is a bug where the game still spawns a head of intelligence and captain of industry despite there already being one of each. I'd like to ask for help with this, and understanding the difference in time zones and schedules, please ping me on Discord or email me to discuss possible solutions. I've checked the characters file, and can't find any leads, so I suspect it might be in zzz_generic.