Added mod files in the corresponding areas (followed naming convention)
I moved the files from my mod into the right folders, renamed the files of pre-existing PAL files, that had shared ISR and PAL stuff, including the loc of said file from PAL_l_english to PAL_ISR_l_english, and the other file named palestine into israel, considering it had the israeli ideas. I did that to make my files which are for palestine to be named as such. +added the portraits of the leaders/generals in gfx/interface/PAL/leaders.