Update focus bypass for "Restore the Liechtenstein Protectorate"
The focus was bypassing immediately due to it checking if Austria was a puppet of itself or in a faction with itself. Since it comes after Austria becomes a puppet of Germany and joins the Axis, that counted. Changed it to checking if Liechtenstein is a puppet of or in a faction with Austria.