Fixed Uruguay starting Navy
Set the proper armaments for Uruguays starting boat.
Historical data shows the ship was build in germany and armed with guns from Krupp, Skoda and Armstrong.
2x 120/45 Krupp QF, which exist word for word in game as Light Battery 1
4x 76/40 EOC N, which is a 76mmx40 gun from Elswick Ordnance a part of Armstrong Whitworth.
6x 37/40 Vickers
No good way to represent those, which means they get a basic AA to show a simple gun.
2x 450mm Torpedo Tube, which is also word for word the same as Torpedo Tube 1
It had a coal powered engine, which had 6k horsepowers, but the weakest i can give it is 35k.
Remaining data is inconsequential to the game.

Now Baris can leave me alone with this one thing...