Zero - Error
Fixes the issue of not gaining A6M Zero as a reward early. - due to being locked behind an airframe you did not have, this creates the aircraft once u have the airframe.

Could be locked by a available i commented out thinking it might be a concern.

Giving any sort of module allows that module to be put on any other airframe due to already exisiting - but this creates other issues of if the reward is worth it.