NEN - Nenetsia
Figures of the 1943 Mandalada rebellion. Sergei Nogo and Vasily Laptander were its leaders. Malevada, the new party name, is described in the Mandalada's Russian Wikipedia article as a peaceful gathering of Nenets, while Mandalada is armed.
https://ru.wikipedia.org/wiki/%D0%9C%D0%B0%D0%BD%D0%B4%D0%B0%D0%BB%D0%B0%D0%B4%D0%B0

Fixed the generic advisors so they don't clash with the generals' advisor traits.
Tyko Vylka made into a character and the Communist leader.