Seychelles
Could not find a leader for the democratic party, the Seychelles Taxpayers and Producers Association, so I went with Marcel Lemarchand, who won a seat for them in the 1948 elections. There does exist the League of Coloured Peoples but if I recall my research a while back correctly it was not specific to Seychelles.