Dominica
Replaced the chief secretary of the Windward Islands Louis Cools-Lartigue with one of the founders of the first party, Phyllis Shand Allfrey of the Dominica Labour Party (founded 1955 with Emmanuel Christopher Loblack though he didn't last). Due to the lack of female generic leader portraits I used one of the Kaiserreich operative ones as a placeholder until I figure out what to do or make one of probably awful quality.